refactor: reorganize packages for improved readability and maintainability
