+++ src/main/java/dev/jhonatta/BookManager/infrastructure/dtos/category/CategoryDTO.java
@@ -1,4 +1,4 @@
-package dev.jhonatta.BookManager.infrastructure.dtos;
+package dev.jhonatta.BookManager.infrastructure.dtos.category;
 
 public record CategoryDTO (Long id,
                            String name){
